bazel: rename go_build_tags extension to dd_go_test
