+++ bazel/rules/dd_go_test/BUILD.bazel
@@ -20,7 +20,7 @@ copy_file(
 go_library(
     name = "gazelle_extension",
     srcs = [":gazelle_extension_src"],
-    importpath = "github.com/DataDog/datadog-agent/bazel/rules/go_build_tags",
+    importpath = "github.com/DataDog/datadog-agent/bazel/rules/dd_go_test",
     visibility = ["//visibility:public"],
     deps = [
         "@gazelle//config",
